QA for Lab 3
1b
added more to the print statements
also fixed the double point precision error in the sum of the array.

1e
part 1 added the defination of multithreading.
part 2 added more to the print statements. 
Also I noted an error in the question "Using Threading library, write a Python program to calculate the factorial of number 21 and square root of number 91.
For the second, you only wrote in the sqroot not its factorial...the problem is also that math.factorial can only take in integers and there does not seem to be a similar function in cmath either.
could perhaps write in a quick loop to manually calc factorial? like:
for i in range(1,x+1):
               factorial=factorial*i